Added base functionality of spawning/despawning players for every already logged in client, fixed few minor bugs and moved thread dispatching direct to Net_MessageInterpreter to handle calls on maint thread

+++ Assets/Scripts/Test Script/Player_Dummy.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 public class Player_Dummy : MonoBehaviour
 {
-    public ushort _pDummyId;
+    public uint _pDummyId;
     public float rotationSpeed = 5f;
     public float movementspeed = 5f;
     private Quaternion targetRotation = Quaternion.identity;
